Purely toroidal case now seems to work
-Various small bugfixes, now toroidal case seems to work
-Included plotter

+++ as_full/sim.h
@@ -24,7 +24,7 @@ extern int thNum;
 extern double dt;
 extern int tNum;
 extern int plotSteps;
-extern int thtd;
+extern double thtd;
 //physical values of functions defining the magnetic field.
 extern double **B;
 extern double **A;
